add catch when null value become from front

+++ Diplom/src/main/java/com/example/diplom/controllers/BucketController.java
@@ -4,6 +4,7 @@ import com.example.diplom.domain.UserM;
 import com.example.diplom.dto.BucketDTO;
 import com.example.diplom.dto.UserNotificationDTO;
 import com.example.diplom.service.BucketService;
+import com.example.diplom.service.ProductService;
 import com.example.diplom.service.UserNotificationService;
 import com.example.diplom.service.UserService;
 import org.springframework.stereotype.Controller;
@@ -18,12 +19,13 @@ import java.util.List;
 public class BucketController {
     private final BucketService bucketService;
     private final UserService userService;
-
+    private final ProductService productService;
     private final UserNotificationService userNotificationService;
 
-    public BucketController(BucketService bucketService, UserService userService, UserNotificationService userNotificationService) {
+    public BucketController(BucketService bucketService, UserService userService, ProductService productService, UserNotificationService userNotificationService) {
         this.bucketService = bucketService;
         this.userService = userService;
+        this.productService = productService;
         this.userNotificationService = userNotificationService;
     }
 
@@ -45,12 +47,18 @@ public class BucketController {
 
     @GetMapping("/my/bucket/{id}/remove")
     public String removeFromBucket(Model model, Principal principal, @PathVariable Long id) {
-
         UserM userM = userService.findByName(principal.getName());
         if (userM == null) {
             throw new RuntimeException("User - " + principal.getName() + " not found");
         }
+
+        if(productService.findProductById(id) == null){
+            model.addAttribute("errorMessage", "Product not found");
+            return "error";
+        }
+
         BucketDTO bucketDTO = bucketService.removeProduct(userM.getBucket(),id,principal.getName());
+
         model.addAttribute("bucket", bucketDTO);
 
         return "redirect:/my/bucket";
@@ -71,14 +79,26 @@ public class BucketController {
 
     @GetMapping("/my/bucket/{id}/increase")
     public String increase(Model model, Principal principal, @PathVariable Long id){
+        if(productService.findProductById(id) == null){
+            model.addAttribute("errorMessage", "Product not found");
+            return "error";
+        }
 
-        bucketService.amountIncrease(userService.findByName(principal.getName()).getBucket(),id);
-        BucketDTO bucketDTO = bucketService.getBucketByUser(principal.getName());
-        model.addAttribute("bucket", bucketDTO);
+        if (principal == null) {
+            model.addAttribute("bucket", new BucketDTO());
+        } else {
+            bucketService.amountIncrease(userService.findByName(principal.getName()).getBucket(),id);
+            BucketDTO bucketDTO = bucketService.getBucketByUser(principal.getName());
+            model.addAttribute("bucket", bucketDTO);
+        }
         return "redirect:/my/bucket";
     }
     @GetMapping("/my/bucket/{id}/decrease")
     public String decrease(Model model, Principal principal, @PathVariable Long id){
+        if(productService.findProductById(id) == null){
+            model.addAttribute("errorMessage", "Product not found");
+            return "error";
+        }
         bucketService.amountDecrease(userService.findByName(principal.getName()).getBucket(),id);
         BucketDTO bucketDTO = bucketService.getBucketByUser(principal.getName());
         model.addAttribute("bucket", bucketDTO);
